+++ lib/runtime/src/cache.rs
@@ -103,7 +103,12 @@ impl Cache for FileSystemCache {
 
         let serialized_cache = Artifact::deserialize(&mmap[..])?;
         unsafe {
-            wasmer_runtime_core::load_cache_with(serialized_cache, &super::default_compiler())
+            wasmer_runtime_core::load_cache_with(
+                serialized_cache,
+                super::compiler_for_backend(backend)
+                    .ok_or_else(|| CacheError::UnsupportedBackend(backend))?
+                    .as_ref(),
+            )
         }
     }